+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Head
+
+- Added: `Steeler` webcrawler verifier.
+
 # 1.1.0 - 2016-10-24
 
 - Added: `Grapeshot` webcrawler verifier.